+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator/test/Providers/ModelProviders/TestData/ModelCustomizationTests/CanCustomizeSpecBaseModelToSystemType/MockInputModel.cs
@@ -1,0 +1,10 @@
+#nullable disable
+
+using System;
+
+namespace Sample.Models
+{
+    public partial class MockInputModel : Exception
+    {
+    }
+}